+++ src/classRun.xml
@@ -1,0 +1,9 @@
+<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd" >
+<suite name="Class Run">
+    <test name="test run">
+        <classes>
+            <class name="tests.US001_Tc01"/>
+            <class name="tests.US001_Tc02"/>
+        </classes>
+    </test>
+</suite>
